refactor(artifacts): colocate Mooncake environment helper
